Added sprites

+++ Obstruction.java
@@ -7,12 +7,12 @@ import java.awt.Rectangle;
  */
 abstract class Obstruction {
   //Coordinates of this obstruction
-  public double xPosition, yPosition;
+  public int xPosition, yPosition;
   public Rectangle boundingBox;
   int height=32;
   int width=32;
   
-  Obstruction(double x, double y) {
+  Obstruction(int x, int y) {
     this.xPosition = x;
     this.yPosition = y;
     this.boundingBox = new Rectangle((int)(this.xPosition), (int)(this.yPosition), width, height);
@@ -23,7 +23,7 @@ abstract class Obstruction {
    * This method returns this object's X-position.
    * @return The X-position of this object.
    */
-  public double getX() {
+  public int getX() {
     return this.xPosition;
   }
 
@@ -32,7 +32,7 @@ abstract class Obstruction {
    * This method sets the X-position of this object.
    * @param The value to set this object's X-position to.
    */
-  public void setX(double newX) {
+  public void setX(int newX) {
     this.xPosition = newX;
   }
 
@@ -41,7 +41,7 @@ abstract class Obstruction {
    * This method returns this object's Y-position.
    * @return The Y-position of this object.
    */
-  public double getY() {
+  public int getY() {
     return this.yPosition;
   }
 
@@ -50,7 +50,7 @@ abstract class Obstruction {
    * This method sets the Y-position of this object.
    * @param The value to set this object's Y-position to.
    */
-  public void setY(double newY) {
+  public void setY(int newY) {
     this.yPosition = newY;
   }
